Add: validation for fs

+++ backend/src/controller/image.controller.js
@@ -7,12 +7,20 @@ export const convertImageController = async (req,res)=>{
     const file = req.file;
     const {Format} =req.body;
 
+    
+
     if(!file || !Format){
         return res.status(400).json({message:"File and format are required"})
     }
 
     try {
         const outputPath = await convertImageFormat(file.path, Format);
+
+        if(!fs.existsSync(outputPath)){
+            return res.status(500).json({message:"Conversion failed : Output not found"})
+        }
+
+        
         res.download(outputPath, `converted.${Format}`, (error)=>{
             if(!error){
                 fs.unlinkSync(file.path)
@@ -21,6 +29,8 @@ export const convertImageController = async (req,res)=>{
                 console.log(error,  "Download failed");
             }
         });
+
+        
 
     } catch (error) {
         console.log("error: conversion failed", error.message);
